move document type enum
