+++ pkg/models/plan_document.go
@@ -56,3 +56,15 @@ func (p PlanDocument) GetModifiedBy() *string {
 func (p PlanDocument) GetCreatedBy() string {
 	return p.CreatedBy
 }
+
+// DocumentType is an enum that represents the type of document
+type DocumentType string
+
+// These constants represent the different values of DocumentType
+const (
+	DocumentTypeConceptPaper   DocumentType = "CONCEPT_PAPER"
+	DocumentTypePolicyPaper    DocumentType = "POLICY_PAPER"
+	DocumentTypeICIPDraft      DocumentType = "ICIP_DRAFT"
+	DocumentTypeMarketResearch DocumentType = "MARKET_RESEARCH"
+	DocumentTypeOther          DocumentType = "OTHER"
+)
